requirements.txt
alembic==1.4.2
attrs==19.3.0
autopep8==1.5.2
bokeh==2.0.1
boto3==1.13.18
botocore==1.16.18
branca==0.4.0
Brotli==1.0.7
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
click-plugins==1.1.1
cligj==0.5.0
dash==1.12.0
dash-bootstrap-components==0.10.0
dash-core-components==1.10.0
dash-dangerously-set-inner-html==0.0.2
dash-html-components==1.0.3
dash-renderer==1.4.1
dash-table==4.7.0
docutils==0.15.2
Flask==1.1.1
Flask-Compress==1.5.0
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
folium==0.10.1
future==0.18.2
geopandas==0.7.0
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.10.0
joblib==0.16.0
Mako==1.1.2
MarkupSafe==1.1.1
munch==2.5.0
packaging==20.3
pandas==1.0.3
Pillow==7.0.0
plotly==4.8.0
pycodestyle==2.6.0
pyparsing==2.4.6
pyproj==2.6.1.post1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
retrying==1.3.3
s3transfer==0.3.3
scikit-learn==0.23.2
scipy
Shapely==1.7.0
six==1.14.0
sklearn==0.0
SQLAlchemy==1.3.15
threadpoolctl==2.1.0
tornado==6.0.4
typing-extensions==3.7.4.1
urllib3==1.25.8
Werkzeug==0.16.1
xlrd==1.2.0
pyOpenSSL==19.1.0
